+++ scripts/provision/provision-conta.sh
@@ -1,0 +1,20 @@
+#!/usr/bin/env bash
+
+set -e
+set -o pipefail
+set -x
+
+RED='\033[0;31m'
+GREEN='\033[0;32m'
+YELLOW='\033[0;33m'
+NC='\033[0m' # No Color
+
+echo -e "${YELLOW} Starting script provision-conta.sh... ${NC}"
+
+echo -e "${GREEN} Starting conta... ${NC}"
+docker-compose up -d conta
+
+echo -e "${GREEN} Running migrations...${NC}"
+docker exec -i ortopedica.conta python manage.py migrate
+
+echo -e "${YELLOW} Finishing script provision-conta.sh... ${NC}"