+++ scripts/provision/provision.sh
@@ -1,5 +1,3 @@
-#!/usr/bin/env bash
-
 set -e
 set -o pipefail
 set -x
@@ -26,6 +24,8 @@ echo -e "MySQL ready"
 
 echo -e "${GREEN}Creating databases and users...${NC}"
 docker exec -i ortopedica.mysql mysql -uroot -ppassword mysql < ./scripts/provision/provision.sql
+
+./scripts/provision/provision-conta.sh
 
 docker image prune -f